DEV-607 fixed raw_value integer->float error

+++ lib/buzzn/schemas/transactions/admin/reading/create.rb
@@ -2,7 +2,7 @@ require './app/models/reading/single.rb'
 require_relative '../reading'
 
 Schemas::Transactions::Admin::Reading::Create = Schemas::Support.Form do
-  required(:raw_value).filled(:bigint?)
+  required(:raw_value).filled(:float?)
   required(:unit).value(included_in?: Reading::Single.units.values)
   required(:reason).value(included_in?: Reading::Single.reasons.values)
   required(:read_by).value(included_in?: Reading::Single.read_by.values)
